+++ src/rm_filescan.cc
@@ -1,41 +1,41 @@
-#include <iostream>
-#include "rm.h"
-using namespace std;
+// #include <iostream>
+// #include "rm.h"
+// using namespace std;
 
 
-RM_FileScan::RM_FileScan()
-{
-  pred = NULL;
-  prmh = NULL;
-  current = RID(1,-1);
-  bopen = False;
-}
+// RM_FileScan::RM_FileScan()
+// {
+//   pred = NULL;
+//   prmh = NULL;
+//   current = RID(1,-1);
+//   bopen = False;
+// }
 
-RM_FileScan::~RM_FileScan()
-{
-}
+// RM_FileScan::~RM_FileScan()
+// {
+// }
 
-RC RM_FileScan::OpenScan  (const RM_FileHandle &fileHandle,
-                AttrType   attrType,
-                int        attrLength,
-                int        attrOffset,
-                CompOp     compOp,
-                void       *value,
-                ClientHint pinHint = NO_HINT) // Initialize a file scan
-{
+// RC RM_FileScan::OpenScan  (const RM_FileHandle &fileHandle,
+//                 AttrType   attrType,
+//                 int        attrLength,
+//                 int        attrOffset,
+//                 CompOp     compOp,
+//                 void       *value,
+//                 ClientHint pinHint = NO_HINT) // Initialize a file scan
+// {
 
     
-}
-RC RM_FileScan::GetNextRec(RM_Record &rec);               // Get next matching record
-{
+// }
+// RC RM_FileScan::GetNextRec(RM_Record &rec);               // Get next matching record
+// {
 
 
 
-}
-RC RM_FileScan::CloseScan ();                             // Close the scan
-{
-  if(!bOpen)
-    return RM_FNOTOPEN;
+// }
+// RC RM_FileScan::CloseScan ();                             // Close the scan
+// {
+//   if(!bOpen)
+//     return RM_FNOTOPEN;
 
 
-}
+// }
